Plein de sphaghetti

+++ lib/AX12/AX12.cpp
@@ -34,7 +34,7 @@ int AX12::move(float Position){
 	if(m_endlessMode){
 		setEndlessMode(OFF);
 	}
-	int pos = min(1023,Position/300*1023);
+	int pos = min((float)1023,Position/300*1023);
 	return Dynamixel.move(m_id, pos);
 }
 
@@ -42,7 +42,7 @@ int AX12::moveSpeed(float Position, float speed){
 	if(m_endlessMode){
 		setEndlessMode(OFF);
 	}
-	int pos = min(1023,Position/300*1023);
+	int pos = min((float)1023,Position/300*1023);
 	return Dynamixel.moveSpeed(m_id, pos, speed);
 }
 
@@ -63,12 +63,12 @@ int AX12::turn(int Speed){
 }
 
 int AX12::Nextmove(float Position){
-	int pos = min(1023,Position/300*1023);
+	int pos = min((float)1023,Position/300*1023);
 	return Dynamixel.moveRW(m_id, pos);
 }
 	
 int AX12::NextmoveSpeed(float Position, float Speed){
-	int pos = min(1023, Position/300*1023);
+	int pos = min((float) 1023, Position/300*1023);
 	return Dynamixel.moveSpeedRW(m_id, pos, Speed);
 }
 	
@@ -82,8 +82,8 @@ int AX12::setTempLimit(unsigned char Temperature){
 }
 	
 int AX12::setAngleLimit(float CW, float CCW){
-	int CWLimit = min(1023, CW/300*1023);
-	int CCWLimit = min(1023, CCW/300 *1023);
+	int CWLimit = min((float)1023, CW/300*1023);
+	int CCWLimit = min((float)1023, CCW/300 *1023);
 	return Dynamixel.setAngleLimit(m_id, CWLimit, CCWLimit);
 }
